requirements.txt
appdirs==1.4.4
asgiref==3.2.10
attrs==19.3.0
beautifulsoup4==4.9.1
black==19.10b0
click==7.1.2
Django==3.0.8
django-bootstrap4==2.2.0
django-dramatiq==0.9.1
django-environ==0.4.5
django-js-asset==1.2.2
django-mptt==0.11.0
django-stronghold==0.4.0
dramatiq==1.9.0
h11==0.9.0
Mako==1.1.3
MarkupSafe==1.1.1
pathspec==0.8.0
pika==1.1.0
prometheus-client==0.8.0
psycopg2-binary==2.8.5
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
redis==3.5.3
regex==2020.6.8
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
uvicorn==0.11.5
websockets==8.1
